web-spider/requirements.txt
beautifulsoup4
lxml
pandas
playwright
python-dotenv
requests
tqdm
PyYAML
